Improved HTML/CSS
Modified the buttons so they're actual links and they have and animation when hovering over them. Changed the behaviour of the background position using clamp() and calc(). This was done so the top of the background always stays at the center of the box, no matter the viewport height (Only exception being a weird behaviour that happens when the vh is smaller than 320px, but since the situation of a screen height being so small is never gonna happen, that's perfectly fine. The background-img also works perfectly fine in 4k screens (both in width and height).